# appservice/src/main/resources/static/swagger/swagger-spec.yaml
swagger: '2.0'
info:
  title: Yona app service API
  description: The API of the service backing the Yona app
  version: 1.0.0
produces:
  - application/json
parameters:
  userID:
    name: userID
    in: path
    description: The ID of the user
    required: true
    type: string
  yonaPassword:
    name: Yona-Password
    in: header
    description: The device password
    required: true
    type: string
...
  buddyID:
    name: buddyID
    in: path
    description: The ID of the buddy
    required: true
    type: string
...
paths:
...
  '/users/{userID}/buddies/':
    post:
      summary: Add a new buddy
      description: Add a new buddy to an existing user
      parameters:
        - $ref: "#/parameters/yonaPassword"
        - $ref: "#/parameters/userID"
        - in: body
          name: body
          description: New buddy
          required: true
          schema:
            $ref: '#/definitions/PostPutBuddy'
      tags:
        - Buddy
      responses:
        '201':
          description: The created buddy
          schema:
            $ref: '#/definitions/Buddy'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    get:
      summary: Get all buddies
      description: Fetches all buddies of the user with the given ID
      parameters:
        - $ref: "#/parameters/yonaPassword"
        - $ref: "#/parameters/userID"
      tags:
        - Buddy
      responses:
        '200':
          description: All buddies of this user.
          schema:
            type: array
            items:
              $ref: '#/definitions/Buddy'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/users/{userID}/buddies/{buddyID}':
    get:
      summary: Get a buddy
      description: Fetches the buddy identified by the given ID
      parameters:
        - $ref: "#/parameters/yonaPassword"
        - $ref: "#/parameters/userID"
        - $ref: "#/parameters/buddyID"
      tags:
        - Buddy
      responses:
        '200':
          description: The requested buddy.
          schema:
            $ref: '#/definitions/Buddy'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      summary: Delete buddy
      description: Deletes this buddy
      parameters:
        - $ref: "#/parameters/yonaPassword"
        - $ref: "#/parameters/userID"
        - $ref: "#/parameters/buddyID"
      tags:
        - Buddy
      responses:
        '204':
          description: The new buddy is deleted.
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
  Curies:
    type: object
    properties:
      curies:
        type: array
        description: A set of curies (Compact URIs). For Yona, there is just one curie.
        items:
          type: object
          properties:
            name:
              type: string
              description: The curie name. The prefix of a curied link will be used to find the curie by name
              example: yona
            href:
              type: string
              description: The URL corresponding to this curie.
              example: http://dev.yona.nu/rels/{rel}
            templated:
              type: string
              format: boolean
              description: Whether this curie is templated (always true for Yona).
              example: true
          required:
            - name
            - href
            - templated
  SelfLink:
    type: object
    properties:
      self:
        type: object
        required:
          - href
        properties:
          href:
            type: string
            description: Identifies this same resource
    required:
      - self
  EditLink:
    type: object
    properties:
      edit:
        type: object
        required:
          - href
        properties:
          href:
            type: string
            description: Through this link, this resource can be edited or deleted. Note that this link is not always present. Certain resources can never be edited. Some others can only be edited or deleted in a specific state. Unprocessed messages for example cannot be deleted.
...
  WeeklyActivityReportsLink:
    type: object
    properties:
      "yona:weeklyActivityReports":
        type: object
        required:
          - href
        properties:
          href:
            type: string
            description: Refers to the weekly activity reports
  DailyActivityReportsLink:
    type: object
    properties:
      "yona:dailyActivityReports":
        type: object
        required:
          - href
        properties:
          href:
            type: string
            description: Refers to the daily activity reports
...
  BaseUser:
    type: object
    properties:
      firstName:
        type: string
        description: First name.
        example: Richard
      lastName:
        type: string
        description: Last name.
        example: Quin
      mobileNumber:
        type: string
        description: 'Mobile number, unique among all users.'
        example: "+31612345678"
      nickname:
        type: string
        description: Nickname known among buddies.
        example: RQ
    required:
    - firstName
    - lastName
    - mobileNumber
    - nickname
...
  UserWithOnlyPublicData:
    allOf:
      - $ref: '#/definitions/BaseUser'
      - type: object
        properties:
          creationTime:
            type: string
            format: dateTime
            description: The creation time of this user
            example: 2016-04-26T20:08:55.365+0200
          appLastOpenedDate:
            type: string
            format: date
            description: The date the user last opened the Yona app
            example: 2016-04-26
          _links:
            allOf:
              - $ref: '#/definitions/SelfLink'
        required:
          - _links
...
  BuddyStatus:
    type: string
    description: The status of this buddy or buddy connect request
    enum:
    - NOT_REQUESTED
    - REQUESTED
    - ACCEPTED
    - REJECTED
  PostPutBuddyUser:
    allOf:
      - $ref: '#/definitions/BaseUser'
      - type: object
        properties:
          emailAddress:
            type: string
            format: email
            description: E-mail address used to inform the buddy about the new buddy request
            example: richard@quin.net
        required:
          - emailAddress
  BaseBuddy:
    type: object
    properties:
      sendingStatus:
        $ref: '#/definitions/BuddyStatus'
      receivingStatus:
        $ref: '#/definitions/BuddyStatus'
    required:
      - sendingStatus
      - receivingStatus
  PostPutBuddy:
    allOf:
      - $ref: '#/definitions/BaseBuddy'
      - type: object
        properties:
          message:
            type: string
            description: A message to the buddy
            example: Hi there, would you want to become my buddy?
          _embedded:
            type: object
            properties:
              "yona:user":
                  $ref: "#/definitions/PostPutBuddyUser"
            required:
              - "yona:user"
        required:
          - _embedded
  Buddy:
    allOf:
      - $ref: '#/definitions/BaseBuddy'
      - type: object
        properties:
          lastStatusChangeTime:
            type: string
            format: dateTime
            description: The time of the last status change (request/accept/reject) of this buddy
            example: 2016-02-23T21:28:58.556+0000
          lastMonitoredActivityDate:
            type: string
            format: date
            description: The date of the last monitored activity of this buddy
            example: 2016-04-26
          _embedded:
            type: object
            properties:
              "yona:user":
                  $ref: "#/definitions/UserWithOnlyPublicData"
            required:
              - "yona:user"
          _links:
            allOf:
              - $ref: '#/definitions/Curies'
              - $ref: '#/definitions/SelfLink'
              - $ref: '#/definitions/EditLink'
              - $ref: '#/definitions/WeeklyActivityReportsLink'
              - $ref: '#/definitions/DailyActivityReportsLink'
        required:
          - lastStatusChangeTime
          - _embedded
          - _links
...
  Error:
    type: object
    properties:
      code:
        type: string
        description: The error code
        example: error.user.mobile.number.invalid
      message:
        type: string
        description: A localized error message
        example: The mobile number ''+31 6 12345678'' is invalid. It must start with a + sign, with no spaces between the digits
      correlationId:
        type: string
        description: ID that correlates the server log messages to this error
        example: 463ac35c9f6413ad48485a3953bb6124
...